+++ lib/index.js
@@ -3,9 +3,10 @@
  * @typedef {import('unist').Position} UnistPosition
  * @typedef {import('vfile-message').VFileMessage} VFileMessage
  * @typedef {import('vscode-languageserver').Connection} Connection
- * @typedef {import('vscode-languageserver-textdocument').TextDocument} TextDocument
- * @typedef {import('vscode-languageserver').TextDocuments<TextDocument>} TextDocuments
- * @typedef {import('..').Options} Options
+ * @typedef {Pick<
+ *   import('unified-engine').Options,
+ *   'ignoreName' | 'packageField' | 'pluginPrefix' | 'plugins' | 'rcName'
+ * >} Options
  */
 
 import {PassThrough} from 'node:stream'
@@ -15,13 +16,17 @@ import {unified} from 'unified'
 import {engine} from 'unified-engine'
 import {VFile} from 'vfile'
 import {
+  createConnection,
   Diagnostic,
   DiagnosticSeverity,
   Position,
+  ProposedFeatures,
   Range,
+  TextDocuments,
   TextDocumentSyncKind,
   TextEdit
 } from 'vscode-languageserver/node.js'
+import {TextDocument} from 'vscode-languageserver-textdocument'
 
 /**
  * Convert a unist point to a language server protocol position.
@@ -117,7 +122,7 @@ function lspDocumentToVfile(document) {
 
 /**
  * @param {Connection} connection
- * @param {TextDocuments} documents
+ * @param {TextDocuments<TextDocument>} documents
  * @param {Options} options
  */
 export function configureUnifiedLanguageServer(
@@ -233,4 +238,20 @@ export function configureUnifiedLanguageServer(
   connection.onDidChangeWatchedFiles(() => {
     checkDocuments(...documents.all())
   })
+}
+
+/**
+ * Create a language server for a unified ecosystem.
+ *
+ * @param {Options} options
+ *   Configuration for `unified-engine` and the language server.
+ */
+export function createUnifiedLanguageServer(options) {
+  const connection = createConnection(ProposedFeatures.all)
+  const documents = new TextDocuments(TextDocument)
+
+  configureUnifiedLanguageServer(connection, documents, options)
+
+  documents.listen(connection)
+  connection.listen()
 }
